#5 - add Canvas metadata to YouTube cap. documents
TL colleagues with Cavas expertise recommend Document objects for YouTube captions should link to the Canvas page on which they were embedded to keep students in the learning context.  Maizey project colleagues say linking to YouTube is fine.  I'm adding extra metadata here for Canvas to show what the links back to Canvas could look like.  They couldn't make the player start at a specific time, but maybe the AI could tell the user the timestamp for the section of video that contains the caption of interest.