#12 - consolidated `canvasapi` imports
Only the PDF module seems to be suitable for importing as needed, because not every Canvas course will contain PDF files that need to be indexed.  I don't see a great advantage for it to not be imported from the very outset, though.  However, I'm trying to refrain from changing anything not related to completing my work.  Consolidating `canvasapi` imports made sense because most of this code needs to work with Canvas; removing the logic around its import simplifies code.